+++ components/colors/src/create-hsl-json.js
@@ -11,7 +11,7 @@ Object.keys(props).forEach((color) => {
     propsHsl[color] = {};
 
     Object.keys(props[color]).forEach((key) => {
-        propsHsl[color][key + "-hsl"] = hexToHsl(props[color][key]);
+        propsHsl[color][key] = hexToHsl(props[color][key]);
     });
 });